# fix: Updated PolygonComponent.containsPoint to account for concave polygons