@davidjan3 2024-01-14
Just stress-tested the performance within my current project.
Calling `containsPoint(vertex)` for 4 vertices on 10,000 randomly generated polygons with 10-24 vertices each:
Old method avg: ~370ms
New method avg: ~530ms

So it takes about 43% longer on average.

However, by replacing 
```dart
final edge = getEdge(i, vertices: vertices);
```
with
```dart
final from = vertices[i];
final to = vertices[(i + 1) % vertices.length];
``` 

It goes down to ~350ms and still survives every test. 
Though this also speeds up the old version to ~320ms.

It's important to note that in my test, the old function returns false almost all the time. And looking at the code, returning false can happen much faster than checking all edges before returning true.