@spydon 2024-01-15
@davidjan3 can you try with a bit more vertices, like 10?