@davidjan3 2024-01-15
All polygons had 10-24 vertices. My test just ran the `.containsPoint(vertex)` method 4 times per polygon, to see if all 4 corners of a rectangle lie within it. I can share my code sample after work.