@spydon 2024-01-15
> Otherwise I'd just add a couple tests and hope for approval.

Sounds good, seems like a reasonable performance trade-off.