@davidjan3 2024-01-15
@spydon So would you prefer keeping the `getEdge()` function for consistency, or should I push this?:


> However, by replacing
> 
> ```dart
> final edge = getEdge(i, vertices: vertices);
> ```
> 
> with
> 
> ```dart
> final from = vertices[i];
> final to = vertices[(i + 1) % vertices.length];
> ```
> 
> It goes down to ~350ms and still survives every test.