@davidjan3 2024-01-16
@spydon Added said optimization, and a little test :)